refactor: keep shared desktop widgets bounded
