+++ packages/ssrvpn_shared/lib/widgets/ssrvpn_subscription_view.dart
@@ -8,6 +8,8 @@ import '../utils/node_country_policy.dart';
 import 'ssrvpn_app_surface.dart';
 
 part 'ssrvpn_subscription_header.dart';
+part 'ssrvpn_subscription_connection_card.dart';
+part 'ssrvpn_subscription_add_card.dart';
 
 enum SsrvpnSubscriptionConnectionStatus {
   disconnected,
@@ -219,227 +221,6 @@ class _SsrvpnSubscriptionViewState extends State<SsrvpnSubscriptionView> {
             ),
           );
         },
-      ),
-    );
-  }
-}
-
-class _SubscriptionConnectionCard extends StatelessWidget {
-  const _SubscriptionConnectionCard({
-    required this.status,
-    required this.currentNodeName,
-  });
-
-  final SsrvpnSubscriptionConnectionStatus status;
-  final String? currentNodeName;
-
-  @override
-  Widget build(BuildContext context) {
-    final normalizedNodeName = currentNodeName?.trim();
-    final hasCurrentNode =
-        status == SsrvpnSubscriptionConnectionStatus.connected &&
-            normalizedNodeName != null &&
-            normalizedNodeName.isNotEmpty;
-    final (label, color, icon) = switch (status) {
-      SsrvpnSubscriptionConnectionStatus.disconnected => (
-          '未连接',
-          SsrvpnUiTokens.textSecondary,
-          Icons.link_off_rounded,
-        ),
-      SsrvpnSubscriptionConnectionStatus.connecting => (
-          '正在连接',
-          SsrvpnUiTokens.warning,
-          Icons.sync_rounded,
-        ),
-      SsrvpnSubscriptionConnectionStatus.connected => (
-          '已连接',
-          SsrvpnUiTokens.success,
-          Icons.link_rounded,
-        ),
-    };
-    final nodeLabel = hasCurrentNode ? normalizedNodeName : null;
-    final semanticsLabel =
-        nodeLabel == null ? '连接状态：$label' : '连接状态：$label。当前节点：$nodeLabel';
-
-    return Semantics(
-      key: const Key('ssrvpn-subscription-status'),
-      container: true,
-      label: semanticsLabel,
-      child: ExcludeSemantics(
-        child: SsrvpnSurfaceCard(
-          padding: const EdgeInsets.symmetric(horizontal: 18, vertical: 15),
-          child: Row(
-            children: [
-              Icon(icon, color: color, size: 22),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      label,
-                      style: TextStyle(
-                        color: color,
-                        fontSize: 15,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
-                    if (nodeLabel != null) ...[
-                      const SizedBox(height: 3),
-                      Row(
-                        children: [
-                          const Text(
-                            '当前节点',
-                            style: TextStyle(
-                              color: SsrvpnUiTokens.textSecondary,
-                              fontSize: 12,
-                            ),
-                          ),
-                          const SizedBox(width: 8),
-                          Expanded(
-                            child: Tooltip(
-                              message: nodeLabel,
-                              child: Text(
-                                compactNodeDisplayName(nodeLabel),
-                                maxLines: 1,
-                                overflow: TextOverflow.ellipsis,
-                                style: const TextStyle(
-                                  color: SsrvpnUiTokens.textPrimary,
-                                  fontSize: 13,
-                                  fontWeight: FontWeight.w600,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _SubscriptionAddCard extends StatelessWidget {
-  const _SubscriptionAddCard({
-    required this.urlController,
-    required this.inputFocusNode,
-    required this.addActionKey,
-    required this.isAdding,
-    required this.isBusy,
-    required this.onAdd,
-  });
-
-  final TextEditingController urlController;
-  final FocusNode inputFocusNode;
-  final GlobalKey addActionKey;
-  final bool isAdding;
-  final bool isBusy;
-  final VoidCallback onAdd;
-
-  @override
-  Widget build(BuildContext context) {
-    return SsrvpnSurfaceCard(
-      padding: const EdgeInsets.all(22),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          const Row(
-            children: [
-              Icon(
-                Icons.add_circle_outline_rounded,
-                color: SsrvpnUiTokens.accent,
-                size: 24,
-              ),
-              SizedBox(width: 10),
-              Expanded(
-                child: Text(
-                  '添加订阅',
-                  style: TextStyle(
-                    color: SsrvpnUiTokens.textPrimary,
-                    fontSize: 19,
-                    fontWeight: FontWeight.w700,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 18),
-          TextField(
-            key: const Key('ssrvpn-subscription-input'),
-            controller: urlController,
-            focusNode: inputFocusNode,
-            enabled: !isBusy,
-            keyboardType: TextInputType.text,
-            textInputAction: TextInputAction.done,
-            autocorrect: false,
-            enableSuggestions: false,
-            scrollPadding: const EdgeInsets.fromLTRB(20, 20, 20, 120),
-            onSubmitted: isBusy ? null : (_) => onAdd(),
-            decoration: InputDecoration(
-              hintText: '粘贴订阅链接或 ssr:// 链接',
-              prefixIcon: const Icon(Icons.link_rounded),
-              filled: true,
-              fillColor: const Color(0xFF181B2A),
-              contentPadding: const EdgeInsets.symmetric(
-                horizontal: 16,
-                vertical: 17,
-              ),
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(18),
-                borderSide: const BorderSide(color: SsrvpnUiTokens.border),
-              ),
-              enabledBorder: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(18),
-                borderSide: const BorderSide(color: SsrvpnUiTokens.border),
-              ),
-            ),
-          ),
-          const SizedBox(height: 16),
-          ConstrainedBox(
-            key: addActionKey,
-            constraints: const BoxConstraints(minHeight: 52),
-            child: SizedBox(
-              width: double.infinity,
-              child: FilledButton(
-                key: const Key('ssrvpn-subscription-add'),
-                onPressed: isBusy ? null : onAdd,
-                style: FilledButton.styleFrom(
-                  backgroundColor: SsrvpnUiTokens.primaryBlue,
-                  foregroundColor: Colors.white,
-                  disabledBackgroundColor:
-                      SsrvpnUiTokens.primaryBlue.withValues(alpha: 0.42),
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(16),
-                  ),
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 16,
-                    vertical: 14,
-                  ),
-                ),
-                child: isAdding
-                    ? const SizedBox(
-                        width: 20,
-                        height: 20,
-                        child: CircularProgressIndicator(
-                          strokeWidth: 2,
-                          color: Colors.white,
-                        ),
-                      )
-                    : const Text(
-                        '添加',
-                        style: TextStyle(
-                            fontSize: 17, fontWeight: FontWeight.w700),
-                      ),
-              ),
-            ),
-          ),
-        ],
       ),
     );
   }
